refactor: improved structure
