+++ ktoolbox/downloader/__init__.py
@@ -1,2 +1,3 @@
 from .base import *
 from .downloader import *
+from .utils import *
