+++ ktoolbox/downloader/downloader.py
@@ -17,8 +17,8 @@ from tqdm import tqdm as std_tqdm
 
 from ktoolbox._enum import RetCodeEnum
 from ktoolbox.configuration import config
-from ktoolbox.downloader import DownloaderRet
-from ktoolbox.utils import filename_from_headers, generate_msg
+from ktoolbox.downloader import DownloaderRet, filename_from_headers
+from ktoolbox.utils import generate_msg
 
 __all__ = ["Downloader"]
